+++ src/main/java/cn/taketoday/web/multipart/AbstractMultipartResolver.java
@@ -21,6 +21,7 @@ package cn.taketoday.web.multipart;
 
 import javax.servlet.http.HttpServletRequest;
 
+import cn.taketoday.web.Constant;
 import cn.taketoday.context.utils.DataSize;
 import lombok.Getter;
 import lombok.Setter;
@@ -37,7 +38,7 @@ public abstract class AbstractMultipartResolver implements MultipartResolver {
 	/*** file upload location */
 	private String location = System.getProperty("java.io.tmpdir");
 
-	private String encoding = "UTF-8";
+	private String encoding = Constant.DEFAULT_ENCODING;
 
 	/**
 	 * Maximum size of a single uploaded file.