+++ CHANGELOG.md
@@ -265,3 +265,4 @@ public final class PathVariableAction {
 - :sparkles: feat: @Interceptor exclude feature
 - :bug: fix: applicationContext NullPointerException
 - refactor: refactor DispatcherServlet
+- unify date format